Update V3.0
Added a launcher and more features. Now there are two maps for the original Death Dungeon, and a map creator. Also Paint Simulator  is another game. New maps can be added with the map creator, and if you want to add them to the game officially, just open maps.txt and in the next new line, paste the code generated by the DD_creator.